requirements.txt
flask
flask_sqlalchemy
flask_cors
mysqlclient
python-dotenv
flask-socketio
gevent
gevent-websocket
